fix: delete duplicate index and fix unit test for mjs
