+++ index.mjs
@@ -2,35 +2,35 @@
  * this files is the entry point for rollup to bundle the library
  * it exports all the classes and functions as named exports
  */
-import Batch from "./src/Batch";
-import ClientBuilder from "./src/ClientBuilder";
-import buildClient from "./src/util/buildClients";
-import SharedCredentials from "./src/SharedCredentials";
-import StaticCredentials from "./src/StaticCredentials";
-import Errors from "./src/Errors";
+import Batch from "./src/Batch.js";
+import ClientBuilder from "./src/ClientBuilder.js";
+import buildClient from "./src/util/buildClients.js";
+import SharedCredentials from "./src/SharedCredentials.js";
+import StaticCredentials from "./src/StaticCredentials.js";
+import Errors from "./src/Errors.js";
 
-import LookupUSStreet from "./src/us_street/Lookup";
-import CandidateUSStreet from "./src/us_street/Candidate";
+import LookupUSStreet from "./src/us_street/Lookup.js";
+import CandidateUSStreet from "./src/us_street/Candidate.js";
 
-import LookupUSZipcode from "./src/us_zipcode/Lookup";
-import ResultUSZipcode from "./src/us_zipcode/Result";
+import LookupUSZipcode from "./src/us_zipcode/Lookup.js";
+import ResultUSZipcode from "./src/us_zipcode/Result.js";
 
-import LookupUSAutocompletePro from "./src/us_autocomplete_pro/Lookup";
-import SuggestionUSAutocompletePro from "./src/us_autocomplete_pro/Suggestion";
+import LookupUSAutocompletePro from "./src/us_autocomplete_pro/Lookup.js";
+import SuggestionUSAutocompletePro from "./src/us_autocomplete_pro/Suggestion.js";
 
-import LookupUSExtract from "./src/us_extract/Lookup";
-import ResultUSExtract from "./src/us_extract/Result";
+import LookupUSExtract from "./src/us_extract/Lookup.js";
+import ResultUSExtract from "./src/us_extract/Result.js";
 
-import LookupInternationalStreet from "./src/international_street/Lookup";
-import CandidateInternationalStreet from "./src/international_street/Candidate";
+import LookupInternationalStreet from "./src/international_street/Lookup.js";
+import CandidateInternationalStreet from "./src/international_street/Candidate.js";
 
-import LookupUSReverseGeo from "./src/us_reverse_geo/Lookup";
+import LookupUSReverseGeo from "./src/us_reverse_geo/Lookup.js";
 
-import LookupInternationalAddressAutocomplete from "./src/international_address_autocomplete/Lookup";
-import SuggestionInternationalAddressAutocomplete from "./src/international_address_autocomplete/Suggestion";
+import LookupInternationalAddressAutocomplete from "./src/international_address_autocomplete/Lookup.js";
+import SuggestionInternationalAddressAutocomplete from "./src/international_address_autocomplete/Suggestion.js";
 
-import LookupUSEnrichment from "./src/us_enrichment/Lookup";
-import ResponseUSEnrichment from "./src/us_enrichment/Response";
+import LookupUSEnrichment from "./src/us_enrichment/Lookup.js";
+import ResponseUSEnrichment from "./src/us_enrichment/Response.js";
 
 export const core = {
   Batch,
